# docs/Plan/27-OPENAPI-SPEC.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "dP Relay v5 API",
    "description": "dP Relay v5 (Modification 6 — Firebase exit): device plane, OTP plane, billing/credits, bulk campaigns, contact groups, templates, and the operator admin plane. Webhook deliveries are signed with X-DP-Signature (hex HMAC-SHA256 of the raw body).",
    "version": "5.3.8-alpha.0"
  },
...
  "paths": {
...
    "/v5/admin/apps/{id}/revoke": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v5/admin/apps/{id}/unrevoke": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v5/admin/apps/{id}/rotate-webhook-secret": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v5/admin/apps/{id}/rotate-device-secret": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v5/admin/apps/{id}/webhook": {
      "patch": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v5/admin/apps/{id}/credentials-status": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
...
}