# server/src/routes/admin-apps.ts
/**
 * Admin app-plane routes (M4 pass 3 remainder, PLAN §10). Operator-gated
 * management of the 004 apps registry: register (with optional server-side
 * secret generation), list, revoke/unrevoke, rotate the webhook secret, and
 * update the webhook URL. Also exposes the global SMS kill switch
 * (settings.kill_switch) so the pause lever behind OTP sends is reachable
 * without DB access, plus the operator-facing aggregate metrics, the
 * cross-app campaign oversight listing, and the per-app credentials-status
 * report (M4 pass 3 deferred items).
 *
 * Auth model (single choke point per plan §5): every route gates via
 * requireOperator (OPERATOR_SECRET) — the same guard as admin billing. These
 * routes never authenticate AS an app; they manage apps.
 *
 * Revocation semantics: requireApp already rejects any app with a non-null
 * revoked_at (401 app_revoked), so revoking here cuts off every app-plane
 * route instantly — OTP send/verify, billing, campaigns, groups, templates.
 *
 * Secret handling: appSecret and webhook_secret are return-once at mint/
 * rotation; only SHA-256 hashes are stored. Listing never exposes secrets.
 */
import type { FastifyPluginAsync, FastifyReply } from "fastify";
import { asRecord, asString } from "../services/parse.js";
import { generateDeviceEnrollmentSecret, generateWebhookSecret, newId, sha256Hex } from "../services/crypto.js";

/** List pagination cap (campaigns parity). */
const LIST_MAX = 100;
const LIST_DEFAULT = 50;

/** Rejects with the structured envelope. */
function fail(reply: FastifyReply, code: number, codeName: string, message: string): FastifyReply {
  return reply.code(code).send({ ok: false, error: message, code: codeName });
}

interface AdminAppRow {
  id: string;
  app_id: string;
  name: string;
  webhook_url: string | null;
  rate_max_per_phone: number;
  rate_window_sec: number;
  created_at: number;
  revoked_at: number | null;
  /** STAGE F9 (ISSUE-88): backing company (NULL = operator-provisioned legacy). */
  company_id: string | null;
  company_name: string | null;
}

function publicApp(row: AdminAppRow) {
  return {
    id: row.id,
    appId: row.app_id,
    name: row.name,
    webhookUrl: row.webhook_url,
    rateMaxPerPhone: row.rate_max_per_phone,
    rateWindowSec: row.rate_window_sec,
    createdAt: row.created_at,
    revokedAt: row.revoked_at,
    // STAGE F9: the operator apps list gains a company column — additive to
    // the projection; legacy apps report null (no company).
    company: row.company_id !== null ? { id: row.company_id, name: row.company_name } : null,
  };
}

/** Validates and normalizes a webhook URL (HTTPS-only, register parity). */
function parseWebhookUrl(reply: FastifyReply, raw: unknown): string | null | undefined {
  if (raw === undefined || raw === null) return undefined;
  if (typeof raw !== "string") {
    fail(reply, 400, "invalid_webhook_url", "webhookUrl must be a string");
    return null;
  }
  let parsed: URL;
  try {
    parsed = new URL(raw);
  } catch {
    fail(reply, 400, "invalid_webhook_url", "webhookUrl must be an absolute https:// URL");
    return null;
  }
  if (parsed.protocol !== "https:") {
    fail(reply, 400, "invalid_webhook_url", "webhookUrl must be an absolute https:// URL");
    return null;
  }
  return parsed.toString();
}

/** Campaign-oversight row (operator listing — joined with apps for the public appId). */
interface OversightRow {
  id: string;
  app_id: string;
  app_public_id: string;
  name: string;
  status: string;
  total_recipients: number;
  sent_count: number;
  failed_count: number;
  created_at: number;
  started_at: number | null;
  completed_at: number | null;
}

const adminAppRoutes: FastifyPluginAsync = async (app) => {
  const db = app.db;

  /**
   * Registers an app (operator variant of POST /v5/apps/register). appSecret
   * may be omitted — the server then generates one (min-length policy holds
   * for caller-supplied secrets either way). The response is the ONLY place
   * the appSecret and webhook_secret are ever returned.
   */
  app.post("/v5/admin/apps", { preHandler: [app.requireOperator] }, async (request, reply) => {
    const body = asRecord(request.body) ?? {};
    const appId = asString(body.appId, 64);
    if (appId === null || !/^[A-Za-z0-9_-]{3,64}$/.test(appId)) {
      return fail(reply, 400, "invalid_app_id", "appId must be 3-64 chars of [A-Za-z0-9_-]");
    }
    let appSecret = asString(body.appSecret, 4096);
    if (body.appSecret !== undefined && body.appSecret !== null && typeof body.appSecret !== "string") {
      return fail(reply, 400, "invalid_app_secret", "appSecret must be a string");
    }
    if (appSecret !== null && appSecret.length < 32) {
      return fail(reply, 400, "invalid_app_secret", "appSecret must be at least 32 chars");
    }
    const generated = appSecret === null;
    if (appSecret === null) appSecret = generateWebhookSecret();

    const name =
      typeof body.name === "string" && body.name.trim().length > 0
        ? body.name.trim().slice(0, 128)
        : appId;
    const webhookUrl = parseWebhookUrl(reply, body.webhookUrl);
    if (webhookUrl === null) return;

    const rateMaxPerPhone =
      typeof body.rateMaxPerPhone === "number" && Number.isInteger(body.rateMaxPerPhone)
        ? body.rateMaxPerPhone
        : 3;
    const rateWindowSec =
      typeof body.rateWindowSec === "number" && Number.isInteger(body.rateWindowSec)
        ? body.rateWindowSec
        : 3600;
    if (rateMaxPerPhone < 1 || rateWindowSec < 1) {
      return fail(reply, 400, "invalid_rate_limits", "rateMaxPerPhone and rateWindowSec must be >= 1");
    }

    const rowId = newId();
    const webhookSecret = generateWebhookSecret();
    // STAGE F7 (ISSUE-87): per-app device enrollment secret, shown once here.
    const deviceEnrollmentSecret = generateDeviceEnrollmentSecret();
    try {
      db.prepare(
        "INSERT INTO apps (id, app_id, app_secret_hash, device_enrollment_secret_hash, name, webhook_url, webhook_secret, " +
          "webhook_secret_hash, rate_max_per_phone, rate_window_sec, created_at) " +
          "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, unixepoch())",
      ).run(rowId, appId, sha256Hex(appSecret), sha256Hex(deviceEnrollmentSecret), name, webhookUrl, webhookSecret, sha256Hex(webhookSecret), rateMaxPerPhone, rateWindowSec);
    } catch (err) {
      if (err instanceof Error && err.message.includes("UNIQUE constraint failed: apps.app_id")) {
        return fail(reply, 409, "app_id_exists", "appId already registered");
      }
      throw err;
    }
    app.log.info({ appId }, "app registered via admin plane");
    const response: Record<string, unknown> = {
      ok: true,
      id: rowId,
      appId,
      name,
      webhookUrl,
      webhookSecret,
      appSecretGenerated: generated,
      // Return-once contract: the operator needs the effective secret either
      // way (a generated one exists nowhere else).
      appSecret,
      // STAGE F7: shown once — the raw deviceEnrollmentSecret exists nowhere else.
      deviceEnrollmentSecret,
    };
    return reply.code(201).send(response);
  });

  /** Lists all apps (never secrets), newest first, keyset-paginated. */
  app.get("/v5/admin/apps", { preHandler: [app.requireOperator] }, async (request) => {
    const query = asRecord(request.query) ?? {};
    const limitRaw = query.limit;
    const limitNum =
      typeof limitRaw === "number"
        ? limitRaw
        : typeof limitRaw === "string" && /^\d+$/.test(limitRaw)
          ? Number.parseInt(limitRaw, 10)
          : NaN;
    const limit = Number.isInteger(limitNum) && limitNum >= 1 ? Math.min(limitNum, LIST_MAX) : LIST_DEFAULT;
    const cursorRaw = query.cursor;
    let cursorAt: number | null = null;
    let cursorId: string | null = null;
    if (typeof cursorRaw === "string") {
      const sep = cursorRaw.indexOf(":");
      const at = sep > 0 ? Number.parseInt(cursorRaw.slice(0, sep), 10) : NaN;
      const id = sep > 0 ? cursorRaw.slice(sep + 1) : "";
      if (Number.isInteger(at) && at >= 0 && id.length > 0) {
        cursorAt = at;
        cursorId = id;
      }
    }

    const rows = db
      .prepare(
        "SELECT a.id, a.app_id, a.name, a.webhook_url, a.rate_max_per_phone, a.rate_window_sec, " +
          "a.created_at, a.revoked_at, a.company_id, c.name AS company_name " +
          "FROM apps a LEFT JOIN companies c ON c.id = a.company_id " +
          "WHERE (? IS NULL OR a.created_at < ? OR (a.created_at = ? AND a.id > ?)) " +
          "ORDER BY a.created_at DESC, a.id ASC LIMIT ?",
      )
      .all(cursorAt, cursorAt, cursorAt, cursorId, limit + 1) as AdminAppRow[];
    const hasMore = rows.length > limit;
    const page = hasMore ? rows.slice(0, limit) : rows;
    return {
      ok: true,
      apps: page.map(publicApp),
      nextCursor: hasMore
        ? `${page[page.length - 1]?.created_at ?? 0}:${page[page.length - 1]?.id ?? ""}`
        : null,
    };
  });

  /** Loads one app by internal id, or null (404 at call site). */
  function loadApp(appRowId: string): AdminAppRow | null {
    return (
      (db
        .prepare(
          "SELECT a.id, a.app_id, a.name, a.webhook_url, a.rate_max_per_phone, a.rate_window_sec, " +
            "a.created_at, a.revoked_at, a.company_id, c.name AS company_name " +
            "FROM apps a LEFT JOIN companies c ON c.id = a.company_id WHERE a.id = ?",
        )
        .get(appRowId) as AdminAppRow | undefined) ?? null
    );
  }

  /**
   * Revokes an app: requireApp rejects every subsequent app-plane request
   * with 401 app_revoked. Campaign/recipient rows are untouched — v5 refunds
   * via the cancel route, not revocation.
   */
  app.post("/v5/admin/apps/:id/revoke", { preHandler: [app.requireOperator] }, async (request, reply) => {
    const params = asRecord(request.params) ?? {};
    const appRowId = asString(params.id, 64);
    if (appRowId === null) return fail(reply, 400, "invalid_app_id", "app id is required");
    const row = loadApp(appRowId);
    if (!row) return fail(reply, 404, "admin_app_not_found", "App not found");
    if (row.revoked_at !== null) {
      return fail(reply, 409, "already_revoked", "App is already revoked");
    }
    db.prepare("UPDATE apps SET revoked_at = unixepoch() WHERE id = ?").run(appRowId);
    app.log.info({ appId: row.app_id }, "app revoked via admin plane");
    return { ok: true, appId: row.app_id, revokedAt: Math.floor(Date.now() / 1000) };
  });

  /** Un-revokes an app (restores app-plane access immediately). */
  app.post("/v5/admin/apps/:id/unrevoke", { preHandler: [app.requireOperator] }, async (request, reply) => {
    const params = asRecord(request.params) ?? {};
    const appRowId = asString(params.id, 64);
    if (appRowId === null) return fail(reply, 400, "invalid_app_id", "app id is required");
    const row = loadApp(appRowId);
    if (!row) return fail(reply, 404, "admin_app_not_found", "App not found");
    if (row.revoked_at === null) {
      return fail(reply, 409, "not_revoked", "App is not revoked");
    }
    db.prepare("UPDATE apps SET revoked_at = NULL WHERE id = ?").run(appRowId);
    app.log.info({ appId: row.app_id }, "app un-revoked via admin plane");
    return { ok: true, appId: row.app_id };
  });

  /**
   * Rotates the webhook secret: mints a new one (returned ONCE here), stores
   * the hash, and stamps webhook_rotated_at. The old secret stops verifying
   * immediately; in-flight dispatch retries signed with it will fail — the
   * audit rows keep their record.
   */
  app.post("/v5/admin/apps/:id/rotate-webhook-secret", { preHandler: [app.requireOperator] }, async (request, reply) => {
    const params = asRecord(request.params) ?? {};
    const appRowId = asString(params.id, 64);
    if (appRowId === null) return fail(reply, 400, "invalid_app_id", "app id is required");
    const row = loadApp(appRowId);
    if (!row) return fail(reply, 404, "admin_app_not_found", "App not found");

    const webhookSecret = generateWebhookSecret();
    db.prepare(
      "UPDATE apps SET webhook_secret = ?, webhook_secret_hash = ?, webhook_rotated_at = unixepoch() WHERE id = ?",
    ).run(webhookSecret, sha256Hex(webhookSecret), appRowId);
    app.log.info({ appId: row.app_id }, "webhook secret rotated via admin plane");
    return { ok: true, appId: row.app_id, webhookSecret, rotatedAt: Math.floor(Date.now() / 1000) };
  });

  /**
   * STAGE F7 (ISSUE-87): mints a NEW per-app device enrollment secret (returned
   * ONCE here) and replaces the stored digest — the previous secret stops
   * matching every future /v5/device/enroll attempt immediately, so a leaked
   * or decommissioned app secret cannot keep binding phones. Apps created
   * before migration 016 (hash NULL) get their first secret from this route.
   */
  app.post("/v5/admin/apps/:id/rotate-device-secret", { preHandler: [app.requireOperator] }, async (request, reply) => {
    const params = asRecord(request.params) ?? {};
    const appRowId = asString(params.id, 64);
    if (appRowId === null) return fail(reply, 400, "invalid_app_id", "app id is required");
    const row = loadApp(appRowId);
    if (!row) return fail(reply, 404, "admin_app_not_found", "App not found");

    const deviceEnrollmentSecret = generateDeviceEnrollmentSecret();
    db.prepare("UPDATE apps SET device_enrollment_secret_hash = ? WHERE id = ?").run(
      sha256Hex(deviceEnrollmentSecret),
      appRowId,
    );
    app.log.info({ appId: row.app_id }, "device enrollment secret rotated via admin plane");
    return {
      ok: true,
      appId: row.app_id,
      deviceEnrollmentSecret,
      rotatedAt: Math.floor(Date.now() / 1000),
    };
  });

  /** Updates the webhook URL (HTTPS-only; empty string clears it). */
  app.patch("/v5/admin/apps/:id/webhook", { preHandler: [app.requireOperator] }, async (request, reply) => {
    const params = asRecord(request.params) ?? {};
    const appRowId = asString(params.id, 64);
    if (appRowId === null) return fail(reply, 400, "invalid_app_id", "app id is required");
    const row = loadApp(appRowId);
    if (!row) return fail(reply, 404, "admin_app_not_found", "App not found");

    const body = asRecord(request.body) ?? {};
    if (body.webhookUrl === undefined) {
      return fail(reply, 400, "invalid_webhook_url", "webhookUrl is required (empty string clears it)");
    }
    if (body.webhookUrl === "" || body.webhookUrl === null) {
      db.prepare("UPDATE apps SET webhook_url = NULL WHERE id = ?").run(appRowId);
      return { ok: true, appId: row.app_id, webhookUrl: null };
    }
    const webhookUrl = parseWebhookUrl(reply, body.webhookUrl);
    if (webhookUrl === null || webhookUrl === undefined) return;
    db.prepare("UPDATE apps SET webhook_url = ? WHERE id = ?").run(webhookUrl, appRowId);
    return { ok: true, appId: row.app_id, webhookUrl };
  });

  /**
   * Credentials-status report (M4 pass 3 deferred item): whether an app's
   * credentials were ever issued and what state they are in now. Apps rows are
   * only created when credentials are minted (operator provisioning, the
   * provisioning route, or the v4 import), so a row's existence IS the answer
   * to "were credentials ever issued" — there is no separate issuance flag to
   * track. Pure read: never touches the row (the dprelay-prod orphaned
   * credential-burn row stays exactly as documented).
   */
  app.get("/v5/admin/apps/:id/credentials-status", { preHandler: [app.requireOperator] }, async (request, reply) => {
    const params = asRecord(request.params) ?? {};
    const appRowId = asString(params.id, 64);
    if (appRowId === null) return fail(reply, 400, "invalid_app_id", "app id is required");
    const row = loadApp(appRowId);
    if (!row) return fail(reply, 404, "admin_app_not_found", "App not found");
    const rotated = (
      db.prepare("SELECT webhook_rotated_at FROM apps WHERE id = ?").get(appRowId) as
        | { webhook_rotated_at: number | null }
        | undefined
    )?.webhook_rotated_at ?? null;
    return {
      ok: true,
      appId: row.app_id,
      credentialsIssued: true,
      revoked: row.revoked_at !== null,
      revokedAt: row.revoked_at,
      webhookConfigured: row.webhook_url !== null,
      webhookRotatedAt: rotated,
      createdAt: row.created_at,
    };
  });

  /**
   * Aggregate metrics (M4 pass 3 deferred item): one operator-facing snapshot
   * across the planes an operator has levers for — apps, users/devices, OTP
   * sessions, bulk campaigns, billing, and webhook delivery health.
   * counts-style COUNT(*) queries (no table scans into JS).
   */
  app.get("/v5/admin/metrics", { preHandler: [app.requireOperator] }, async () => {
    const count = (sql: string): number => (db.prepare(sql).get() as { n: number }).n;
    const now = Math.floor(Date.now() / 1000);
    const dayAgo = now - 86_400;
    return {
      ok: true,
      generatedAt: now,
      apps: {
        total: count("SELECT COUNT(*) AS n FROM apps"),
        revoked: count("SELECT COUNT(*) AS n FROM apps WHERE revoked_at IS NOT NULL"),
      },
      users: {
        total: count("SELECT COUNT(*) AS n FROM users"),
        devices: count("SELECT COUNT(*) AS n FROM devices"),
      },
      otp: {
        sessionsTotal: count("SELECT COUNT(*) AS n FROM otp_sessions"),
        sessionsPending: count("SELECT COUNT(*) AS n FROM otp_sessions WHERE status = 'pending'"),
        sessionsVerified: count("SELECT COUNT(*) AS n FROM otp_sessions WHERE status = 'verified'"),
        sessionsLast24h: count(`SELECT COUNT(*) AS n FROM otp_sessions WHERE created_at >= ${dayAgo}`),
      },
      bulk: {
        campaignsTotal: count("SELECT COUNT(*) AS n FROM bulk_campaigns"),
        campaignsActive: count("SELECT COUNT(*) AS n FROM bulk_campaigns WHERE status IN ('queued', 'sending', 'paused')"),
        recipientsSent: count("SELECT COUNT(*) AS n FROM bulk_recipients WHERE status = 'sent'"),
        recipientsFailed: count("SELECT COUNT(*) AS n FROM bulk_recipients WHERE status = 'failed'"),
        recipientsQueued: count("SELECT COUNT(*) AS n FROM bulk_recipients WHERE status IN ('pending', 'queued')"),
      },
      billing: {
        transactionsPending: count("SELECT COUNT(*) AS n FROM credit_transactions WHERE status = 'pending'"),
        transactionsApproved: count("SELECT COUNT(*) AS n FROM credit_transactions WHERE status = 'approved'"),
        transactionsRejected: count("SELECT COUNT(*) AS n FROM credit_transactions WHERE status = 'rejected'"),
        creditsRows: count("SELECT COUNT(*) AS n FROM app_credits"),
      },
      webhooks: {
        deliveriesLast24h: count(`SELECT COUNT(*) AS n FROM webhook_deliveries WHERE created_at >= ${dayAgo}`),
        deliveredLast24h: count(`SELECT COUNT(*) AS n FROM webhook_deliveries WHERE created_at >= ${dayAgo} AND status = 'delivered'`),
        failedLast24h: count(`SELECT COUNT(*) AS n FROM webhook_deliveries WHERE created_at >= ${dayAgo} AND status = 'failed'`),
      },
    };
  });

  /**
   * Campaign oversight listing (M4 pass 3 deferred item): every campaign
   * across ALL apps, newest first — the operator counterpart to the per-app
   * GET /v5/bulk/campaigns (which scopes to the calling app). Same keyset
   * cursor contract; optional status filter, same enum as the app plane.
   */
  app.get("/v5/admin/campaigns", { preHandler: [app.requireOperator] }, async (request, reply) => {
    const query = asRecord(request.query) ?? {};
    const statusRaw = query.status;
    const status =
      statusRaw === "queued" || statusRaw === "sending" || statusRaw === "paused" ||
      statusRaw === "completed" || statusRaw === "cancelled"
        ? statusRaw
        : null;
    if (statusRaw !== undefined && status === null) {
      return fail(reply, 400, "invalid_status", "status must be one of queued, sending, paused, completed, cancelled");
    }
    const limitRaw = query.limit;
    const limitNum =
      typeof limitRaw === "number"
        ? limitRaw
        : typeof limitRaw === "string" && /^\d+$/.test(limitRaw)
          ? Number.parseInt(limitRaw, 10)
          : NaN;
    const limit = Number.isInteger(limitNum) && limitNum >= 1 ? Math.min(limitNum, LIST_MAX) : LIST_DEFAULT;
    const cursorRaw = query.cursor;
    let cursorAt: number | null = null;
    let cursorId: string | null = null;
    if (typeof cursorRaw === "string") {
      const sep = cursorRaw.indexOf(":");
      const at = sep > 0 ? Number.parseInt(cursorRaw.slice(0, sep), 10) : NaN;
      const id = sep > 0 ? cursorRaw.slice(sep + 1) : "";
      if (Number.isInteger(at) && at >= 0 && id.length > 0) {
        cursorAt = at;
        cursorId = id;
      }
    }

    const rows = db
      .prepare(
        "SELECT c.id, c.app_id, c.name, c.status, c.total_recipients, c.sent_count, c.failed_count, " +
          "c.created_at, c.started_at, c.completed_at, a.app_id AS app_public_id " +
          "FROM bulk_campaigns c JOIN apps a ON a.id = c.app_id " +
          "WHERE (? IS NULL OR c.status = ?) " +
          "AND (? IS NULL OR c.created_at < ? OR (c.created_at = ? AND c.id > ?)) " +
          "ORDER BY c.created_at DESC, c.id ASC LIMIT ?",
      )
      .all(status, status, cursorAt, cursorAt, cursorAt, cursorId, limit + 1) as OversightRow[];
    const hasMore = rows.length > limit;
    const page = hasMore ? rows.slice(0, limit) : rows;
    return {
      ok: true,
      campaigns: page.map((row) => ({
        campaignId: row.id,
        appId: row.app_public_id,
        name: row.name,
        status: row.status,
        totalRecipients: row.total_recipients,
        sentCount: row.sent_count,
        failedCount: row.failed_count,
        queuedCount: Math.max(0, row.total_recipients - row.sent_count - row.failed_count),
        createdAt: row.created_at,
        startedAt: row.started_at,
        completedAt: row.completed_at,
      })),
      nextCursor: hasMore
        ? `${page[page.length - 1]?.created_at ?? 0}:${page[page.length - 1]?.id ?? ""}`
        : null,
    };
  });

  /**
   * Global SMS kill switch (settings.kill_switch, seeded by migration 001):
   * the pause lever behind POST /v5/otp/send — when on, sends reject with
   * 503 sms_paused while verification of already-delivered codes keeps
   * working (routes/otp.ts). `enabled: true` pauses, `false` resumes; the
   * response reports the PREVIOUS state so the caller knows what the flip
   * changed, and no-op flips to the current state never touch the row.
   */
  app.post("/v5/admin/kill-switch", { preHandler: [app.requireOperator] }, async (request, reply) => {
    const body = asRecord(request.body) ?? {};
    if (typeof body.enabled !== "boolean") {
      return fail(reply, 400, "invalid_enabled", "enabled must be a boolean (true pauses SMS, false resumes)");
    }
    const row = db
      .prepare("SELECT value FROM settings WHERE key = 'kill_switch'")
      .get() as { value: string } | undefined;
    if (!row) {
      // 001 seeds this row; a missing row means a foreign/pre-migration DB.
      return fail(reply, 500, "kill_switch_missing", "kill_switch setting is not initialized");
    }
    const previous = row.value === "true";
    if (previous === body.enabled) {
      return { ok: true, previous, enabled: body.enabled, changed: false };
    }
    db
      .prepare("UPDATE settings SET value = ?, updated_at = unixepoch() WHERE key = 'kill_switch'")
      .run(body.enabled ? "true" : "false");
    app.log.info({ previous, enabled: body.enabled }, "SMS kill switch toggled via admin plane");
    return { ok: true, previous, enabled: body.enabled, changed: true };
  });
};

export default adminAppRoutes;